First commit of CellML version of respiration block of Ikeda model.
This was constructed by starting with the Berkley-Madonna version of the model that I received by e-mail from Bernard de Bono, who obtained it from Randy Thomas.  The original code has a comment saying it was implemented by Julie Fontecave. The syntax was converted to the notation used by OpenCell v0.8, and then copied and pasted into OpenCell.

This is just the maths, no variables declared yet, and no units.